Revert "Split `FrameworkHook` into multiple parts (#3561)"
This reverts commit b75fec1e968ba40709762a10dcf0ad278075b5c0.
